Experiment 4 changes to prevent disk space getting full
We now spawn a script to keep clearing the file (to which the dd command outputs to)
at regular intervals. This makes the given file sparse but we still achieve our
objective as the actual space occupied remains constrained within 10 GB.
Choosing an interval of 20 secs has higher impact on slowness than lower values(5,10secs).
We can still go higher(30 secs) but this will also increase the file size intermittently
which we want to avoid.