Revamped experiment4 disk contention
Use dd command to write in a loop. Use flag conv=notrunc, to make inplace modification
(over-write to the same file). We no longer need any cleanup script or rm command to remove the file.
The flag keeps the max file size bounded to the specified file.
The value chosen is 1.6GB for CRDB and 1.8GB for RethinkDB.
Started from 200MB and kept incrementing in steps of 200MB until the disk utilizaton seen
by iostat for given disk was 100%(for most of the time).