Changes for experiment 4
Exp 4 now uses modified dd command that does in-place modification to the output file(overwrite the same file). Having this essentially means we no longer need any cleanup or removal of the file to prevent from disk exhaustion.
Also the file size can be tuned to each DB. The way i did was starting from 200MB and increasing in steps of 200MB until the disk utilizaton seen
by iostat command for given disk(on the target node) was 100%.
We also allocate a fixed sized file(A) on all the nodes even before the DB process is started. Later during the experiment, dd command will keep writing to the same file A. This would prevent any changes in free space and TiDB should not trigger leader re-election(the current theory is that TiDB monitors free space to trigger leader re-election). The same change is replicated for other DBs to keep config consistent.